conv edges instead of mask
