+++ deepstruct/node_map_strategies.py
@@ -96,16 +96,20 @@ class Conv2LayerMapper(NodeMapper):
         input_width = input_shape(output_width, 1)
         input_neurons_count = channels_in * input_width * input_height
         output_neurons_count = channels_out * output_height * output_width
-        # mask = torch.ones((output_neurons_count, input_neurons_count), dtype=torch.bool)
-        # mask[:, :] = False
-        mask = torch.zeros((output_neurons_count, input_neurons_count), dtype=torch.bool)
+        input_neurons = graph.add_vertices(input_neurons_count,
+                                           node_name,
+                                           output_layer_size=output_neurons_count,
+                                           **kwargs
+                                           )
+        output_neurons = [input_neurons[-1] + i + 1 for i in range(output_neurons_count)]
 
         def get_input_neuron(channel: int, row: int, col: int):
-            return int((col * input_height + row) + (channel * input_width * input_height))
+            return int(input_neurons[int((col * input_height + row) + (channel * input_width * input_height))])
 
         def get_output_neuron(channel_out: int, row: int, col: int):
-            return int((col * output_height + row) + (channel_out * output_width * output_height))
+            return int(output_neurons[int((col * output_height + row) + (channel_out * output_width * output_height))])
 
+        edges = []
         for idx_channel_out in range(channels_out):
             for idx_channel_in in range(channels_in):
                 out_col = 0
@@ -115,16 +119,13 @@ class Conv2LayerMapper(NodeMapper):
                     offset_width = -padding[1]
                     while offset_width + size_kernel[1] <= input_width:
                         target = get_output_neuron(idx_channel_out, out_row, out_col)
-                        for col in range(max(0, offset_height), min(offset_height + size_kernel[0], input_height),):
-                            for row in range(max(0, offset_width), min(offset_width + size_kernel[1], input_width),):
+                        for col in range(max(0, offset_height), min(offset_height + size_kernel[0], input_height), ):
+                            for row in range(max(0, offset_width), min(offset_width + size_kernel[1], input_width), ):
                                 source = get_input_neuron(idx_channel_in, row, col)
-                                mask[target, source] = True
+                                edges.append((source, target))
                         offset_width += stride[1]
                         out_row += 1
                     offset_height += stride[0]
                     out_col += 1
-
-        kwargs['mask'] = mask
-        graph.add_vertices(input_neurons_count, node_name, output_layer_size=output_neurons_count, **kwargs)
+        graph.edges_for_name[node_name] = edges
         graph.add_edges(predecessors, node_name)
-
